More tests are passing

+++ src/usecases/DeleteTodoUC.js
@@ -1,13 +1,19 @@
 const UsersRepository = require("../repository/UsersRepository");
 
 class DeleteTodoUC {
-  async execute(request, response) {
+  execute(request, response) {
     const todoId = request.params.id;
     const username = request.header("username");
 
+    if (!UsersRepository.isTodoAlreadyExists(username, todoId)) {
+      return response.status(404);
+    }
+
     UsersRepository.delete(todoId, username);
 
-    return response.json(UsersRepository.getAllTodos(username));
+    const allTodos = UsersRepository.getAllTodos(username);
+
+    return response.status(204).json(allTodos);
   }
 }
